Move unitary logic into separate compile unit
