+++ src/CMakeLists.txt
@@ -19,6 +19,7 @@ ADD_LIBRARY(ql
   arch/cc/codegen_cc.cc
   arch/cc/vcd.cc
   cqasm/cqasm_reader.cc
+  unitary.cc
   )
 
 # support 'include <ql/foo.h>'
